+++ src/restangular.js
@@ -228,6 +228,11 @@ module.provider('Restangular', function() {
                   isCollection, route, Restangular);
             }
             
+            config.fullResponse = _.isUndefined(config.fullResponse) ? false : config.fullResponse;
+            object.setFullResponse = function(full) {
+                config.fullResponse = full;
+            }
+            
             
             
             //Internal values and functions
@@ -452,6 +457,16 @@ module.provider('Restangular', function() {
                   });
                   return restangularizePromise(deferred.promise, this[config.restangularFields.restangularCollection]);
               }
+
+              function resolvePromise(deferred, response, data) {
+                if (config.fullResponse) {
+                  return deferred.resolve(_.extend(response, {
+                    data: data
+                  }));
+                } else {
+                  deferred.resolve(data);
+                }
+              }
               
               
               // Elements
@@ -552,9 +567,9 @@ module.provider('Restangular', function() {
 
                       processedData = _.extend(data, processedData);
                       if (!__this[config.restangularFields.restangularCollection]) {
-                          deferred.resolve(restangularizeCollection(__this, processedData, what));
+                          resolvePromise(deferred, response, restangularizeCollection(__this, processedData, what));
                       } else {
-                          deferred.resolve(restangularizeCollection(null, processedData, __this[config.restangularFields.route]));
+                          resolvePromise(deferred, response, restangularizeCollection(null, processedData, __this[config.restangularFields.route]));
                       }
                   }, function error(response) {
                       config.errorInterceptor(response);
@@ -580,9 +595,9 @@ module.provider('Restangular', function() {
                       var resData = response.data;
                       var elem = config.responseExtractor(resData, operation, route, fetchUrl) || resObj;
                       if (operation === "post" && !__this[config.restangularFields.restangularCollection]) {
-                        deferred.resolve(restangularizeElem(__this, elem, what));
+                        resolvePromise(deferred, response, restangularizeElem(__this, elem, what));
                       } else {
-                        deferred.resolve(restangularizeElem(__this[config.restangularFields.parentResource], elem, __this[config.restangularFields.route]));
+                        resolvePromise(deferred, response, restangularizeElem(__this[config.restangularFields.parentResource], elem, __this[config.restangularFields.route]));
                       }
 
                   };